Close the evidence-layer gap: packaging, CI parity, HRTF sample-rate, examples, bench (P2)
Packaging (audit B14): install/export rules + AmbiTapConfig.cmake make the
README's find_package(AmbiTap) claim true. Both Eigen resolutions export
correctly: a system Eigen is recorded by name and resolved by the config's
find_dependency(Eigen3); a FetchContent Eigen exports as an empty shim the
config forwards to the consumer's Eigen. Locally verified end-to-end for
both paths; a CI job installs to a prefix and builds a downstream consumer.

CI parity: GCC/Clang/AppleClang/MSVC matrix with AMBITAP_WERROR=ON and
example smoke runs, the ASan+UBSan and TSan legs, a SOFA build leg, and the
install/find_package smoke test. Tests/examples/bench build against a strict
warnings interface target (-Wall -Wextra -Wpedantic -Wconversion -Wshadow,
/W4 /permissive-) with zero warnings.

HRTF sample rate (B7): binaural_renderer::prepare(block, sample_rate)
resamples the built-in 44.1 kHz FIRs to the host rate via a windowed-sinc
resampler (math/binaural/resample.h); centroid-scaling regression test.

Examples and benchmarks: examples/encode_rotate_decode (offline end-to-end
path with per-speaker RMS report), examples/binaural_render (orbiting source
written to WAV, exercises the 48 kHz resample path); bench/ measures
us/block per processor per order plus worker-thread rebuild latency,
dependency-free.

Build/repo hygiene: the stale multi-language .clang-format from an unrelated
project is replaced with a config matching the house style plus a
.clang-format-ignore for generated tables (one-time reformat + CI gate left
as an explicit author decision); libmysofa BUILD_TESTS cache clobber removed
(pin-to-SHA left as a TODO — the hash is unreachable from this environment);
GTest prefers a system install via FIND_PACKAGE_ARGS; gtest discovery/test
timeouts; Eigen fetch uses MakeAvailable+SOURCE_SUBDIR on CMake >= 3.28
(deprecated Populate only as a pre-3.28 fallback); cmake_minimum_required
3.21 -> 3.24; expanded .gitignore.

Header quality: HRTF tables are inline constexpr (one copy per binary;
generator emits the same); M_PI removed from library headers (portable k_pi
+ shared to_spherical helper in coords.h, reused by allrad); rotation.h no
longer claims the least-squares construction is more robust than
Ivanic-Ruedenberg and documents the Euler convention (positive pitch =
front-down). README rewritten: accurate tree (dsp/, analysis/), RT contract
stated and tied to the tests that enforce it, consuming/options tables,
badges, audit link.

Suite: 99 tests, green under plain, ASan+UBSan, and TSan builds.

+++ tests/test_core.cpp
@@ -47,7 +47,7 @@ TEST(Normalization, Sn3dBasics) {
 TEST(Normalization, N3dRelations) {
     for (int n = 0; n <= max_order; ++n) {
         for (int m = 0; m <= n; ++m) {
-            EXPECT_NEAR(n3d_factor(n, m), sn3d_factor(n, m) * std::sqrt(2.0f * n + 1.0f), 1e-6f);
+            EXPECT_NEAR(n3d_factor(n, m), sn3d_factor(n, m) * std::sqrt(2.0f * static_cast<float>(n) + 1.0f), 1e-6f);
         }
         EXPECT_NEAR(sn3d_to_n3d(n) * n3d_to_sn3d(n), 1.0f, 1e-6f);
     }
